Very modularized + addition to visualize it
Here I took everything to the modules. Before, there were some parts of the upload module in the server.R . Now, everything related to a module is sent to that module.

I also added some changes to the visualize it tab. Mainly, I added theme selection and plot type selection. I also added an actionbutton, so that the a plot is made only after clicking it. This will prevent the app from blocking, e.g. when a datetime column exists, and gives time to first convert that column to a POSIXct object for instance in the tidy it tab.