Corrected row filtering code
This commit is just a correction of some mistakes in the code. The code runs fine now.
A few remarks: Make sure you don't make any mistakes when reusing the IDs. Lowercase and uppercase already make it not work.
Pay attention also to the opening and closing of the brackets... The rendertable was outside the observeevent, so it won't be displayed when the button is clicked....